Improve event generator (#8)
Event generator will automatically add timestamps, and will only include a response if the event type is "attempted" or "submitted".

Also allows passing a map of keys and values (agnostic to the nesting structure of the event) to create an event based on this data. Note: I'm assuming this is a reasonable approach given that the key names will all be distinct based on the spec we make.