Voila introspect initial ui (#9)
* Temp change producing consistent user-ids for events

* Score extraction now parses string into float

* Added batch event processor

* Test cleanup Event Processor

* Event will now generate with timestamps, and include responses only if the action verb is attempted or submitted

* Events can now be produced by giving the generator a map of keys and values to include

* First pass at ui for voila

* Potential extension of the api to allow external events

Create dependencies (called components in this commit) at the start of the
application so that they can be passed in to other parts of the program that
require them from the top. This means we can always pass in dependencies for a
given piece of code and can isolate it more easily when it comes to testing.

The event-processor should modify it's own storage, and when it comes time to
render a learner-representation we should combine the "user entity" with the
event-processors stored "aggregate representation" to be displayed. This would
naturally take place within the learner-representation handler.

* Add get event endpoint

* wip changes for ui demo

* Remove some gubbins, add correctly functioning api-root

Co-authored-by: enanae <ane.espeseth@gmail.com>